+++ nix/materialized/x86_64-linux/project/.plan.nix/eventuo11y-dsl.nix
@@ -19,7 +19,7 @@
       homepage = "";
       url = "";
       synopsis = "DSL for defining eventuo11y fields and selectors";
-      description = "Exposes a DSL for low-boilerplate definition of [eventuo11y](https://hackage.haskell.org/package/eventuo11y) fields and selectors.\n\nSee \"Observe.Event.DSL\" for the core DSL.\n\nSee \"Observe.Event.DSL.Compile\" for the TemplateHaskell code to generate the field and selector types.\n\nSee [Example.hs](https://github.com/shlevy/eventuo11y/tree/v0.5.0.0/Example.hs) for an example.\n\nPackages providing @EventBackend@s should likely also provide extensions to the DSL and generate default renderers.";
+      description = "Exposes a DSL for low-boilerplate definition of [eventuo11y](https://hackage.haskell.org/package/eventuo11y) fields and selectors.\n\nSee \"Observe.Event.DSL\" for the core DSL.\n\nSee \"Observe.Event.DSL.Compile\" for the TemplateHaskell code to generate the field and selector types.\n\nSee [Example.hs](https://github.com/shlevy/eventuo11y/tree/v0.6.0.0/Example.hs) for an example.\n\nPackages providing @EventBackend@s should likely also provide extensions to the DSL and generate default renderers.";
       buildType = "Simple";
       isLocal = true;
       detailLevel = "FullDetails";